requirements.txt
aioredis==1.3.1
anyio==3.6.2
async-timeout==4.0.2
click==8.1.3
colorama==0.4.6
Deprecated==1.2.13
fastapi==0.95.1
fastapi-cache2==0.2.1
fastapi-pagination==0.12.2
h11==0.14.0
hiredis==2.2.3
idna==3.4
importlib-resources==5.12.0
limits==3.4.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pendulum==2.1.2
pydantic==1.10.7
python-dateutil==2.8.2
pytz==2023.3
pytzdata==2020.1
six==1.16.0
slowapi==0.1.8
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
tzdata==2023.3
uvicorn==0.22.0
wrapt==1.15.0
